componentized TransportContorls and switched to stores instead of props
