+++ drum-machine/src/components/PatternTimeline/PatternTimeline.css
@@ -1,10 +1,9 @@
-/* PatternTimeline.css - Fixed Width Design (No Responsive) */
+/* PatternTimeline.css - Grid-focused styles (Transport controls removed) */
 
 /* CSS Custom Properties for easy theming and consistency */
 :root {
   --timeline-bg: #ffffff;
   --timeline-border: #dee2e6;
-  --timeline-controls-bg: #f8f9fa;
   --timeline-sidebar-bg: #f8f9fa;
   --timeline-header-bg: #ffffff;
 
@@ -44,174 +43,6 @@
   /* Uses shared .floating-card class - fixed 1200px width */
   overflow: hidden;
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
-}
-
-/* Controls Bar */
-.timeline-controls {
-  background: var(--timeline-controls-bg);
-  border-bottom: 1px solid var(--timeline-border);
-  padding: 16px 24px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: 16px;
-}
-
-.controls-section {
-  display: flex;
-  align-items: center;
-  gap: 24px;
-}
-
-/* Transport Controls - UPDATED to match mockup */
-.transport-controls {
-  display: flex;
-  gap: 8px;
-}
-
-.transport-btn {
-  border: none;
-  border-radius: 12px;
-  padding: 12px 20px;
-  font-weight: 600;
-  font-size: 14px;
-  cursor: pointer;
-  transition: all var(--transition-fast);
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 6px;
-  box-shadow: var(--shadow-sm);
-  color: white;
-}
-
-.transport-btn--base {
-  background: #6c757d;
-  color: white;
-}
-
-.transport-btn--base:hover {
-  transform: translateY(-1px);
-  box-shadow: var(--shadow-md);
-}
-
-.transport-btn--play {
-  background: #28a745;
-  width: 95px; /* Fixed width to prevent button movement */
-}
-
-.transport-btn--play:hover {
-  background: #218838;
-  transform: translateY(-1px);
-  box-shadow: var(--shadow-md);
-}
-
-.transport-btn--pause {
-  background: #ffc107;
-  color: #212529;
-  width: 95px; /* Fixed width to prevent button movement */
-}
-
-.transport-btn--pause:hover {
-  background: #e0a800;
-  transform: translateY(-1px);
-  box-shadow: var(--shadow-md);
-}
-
-.transport-btn--stop {
-  background: #dc3545;
-}
-
-.transport-btn--stop:hover {
-  background: #c82333;
-  transform: translateY(-1px);
-  box-shadow: var(--shadow-md);
-}
-
-.transport-btn--loop {
-  background: #6f42c1; /* Back to purple */
-}
-
-.transport-btn--loop:hover {
-  background: #5a32a3; /* Back to purple hover */
-  transform: translateY(-1px);
-  box-shadow: var(--shadow-md);
-}
-
-/* Position Display */
-.position-display {
-  display: flex;
-  gap: 8px;
-}
-
-.position-badge {
-  padding: 4px 8px;
-  border-radius: 4px;
-  font-size: 12px;
-  font-weight: 600;
-  color: white;
-}
-
-.position-badge--beat {
-  background: #17a2b8;
-}
-
-.position-badge--tick {
-  background: #6c757d;
-}
-
-/* Snap Control */
-.snap-control {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.snap-checkbox {
-  margin: 0;
-}
-
-.snap-label {
-  margin: 0;
-  font-weight: 500;
-  color: #495057;
-  cursor: pointer;
-}
-
-/* BPM Control - UPDATED to match mockup style */
-.bpm-control {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  background: rgba(91, 192, 235, 0.1);
-  padding: 12px 16px;
-  border-radius: 12px;
-  border: 1px solid rgba(91, 192, 235, 0.2);
-}
-
-.bpm-label {
-  font-weight: 600;
-  color: #5bc0eb;
-  margin: 0;
-  font-size: 14px;
-}
-
-.bpm-slider {
-  width: 120px;
-  margin: 0;
-  accent-color: #5bc0eb;
-}
-
-.bpm-value {
-  background: #5bc0eb;
-  color: white;
-  padding: 4px 8px;
-  border-radius: 8px;
-  font-weight: 600;
-  font-size: 14px;
-  min-width: 48px; /* Fixed width to prevent size changes */
-  text-align: center;
 }
 
 /* Grid Container - Flexbox layout */
@@ -568,83 +399,10 @@
   }
 }
 
-/* Focus states for accessibility */
-.transport-btn:focus,
-.snap-checkbox:focus,
-.bmp-slider:focus {
-  outline: 2px solid var(--playhead-color);
-  outline-offset: 2px;
-}
-
 /* Print styles */
 @media print {
-  .timeline-controls {
-    display: none;
-  }
-
   .pattern-timeline {
     box-shadow: none;
     border: 1px solid #000;
   }
-}
-
-/* Measure Controls - Mockup Style */
-.measure-controls {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.measure-label {
-  color: var(--text-muted, #adb5bd);
-  font-weight: bold;
-  font-size: 14px;
-  margin: 0;
-}
-
-.measure-btn {
-  padding: 4px 8px;
-  border: 1px solid var(--border-color, #dee2e6);
-  background: white;
-  border-radius: 6px;
-  cursor: pointer;
-  font-size: 14px;
-  font-weight: 500;
-  color: var(--text-primary, #333);
-  transition: all 0.15s ease;
-  min-width: 28px;
-  height: 28px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.measure-btn:hover {
-  background: #f8f9fa;
-  border-color: #adb5bd;
-}
-
-.measure-btn:active {
-  background: #e9ecef;
-}
-
-.measure-btn:disabled {
-  background: #f8f9fa;
-  border-color: #e9ecef;
-  color: #adb5bd;
-  cursor: not-allowed;
-}
-
-.measure-count {
-  background: var(--text-secondary, #6c757d);
-  color: white;
-  padding: 4px 8px;
-  border-radius: 6px;
-  font-weight: 600;
-  font-size: 14px;
-  min-width: 28px;
-  height: 28px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
 }
