+++ drum-machine/src/components/TransportControls/TransportControls.css
@@ -1,56 +1,29 @@
-/* Transport Controls Container - replaces Bootstrap utilities */
-.transport-container {
-  background: #f8f9fa;
-  padding: 12px 16px;
-  border-bottom: 1px solid #dee2e6;
-  margin-bottom: 16px;
+/* TransportControls.css - Extracted from PatternTimeline.css */
+
+/* Controls Bar */
+.timeline-controls {
+  background: var(--timeline-controls-bg, #f8f9fa);
+  border-bottom: 1px solid var(--timeline-border, #dee2e6);
+  padding: 16px 24px;
   display: flex;
   justify-content: space-between;
   align-items: center;
+  flex-wrap: wrap;
+  gap: 16px;
 }
 
-/* Left Section - Transport Buttons */
-.transport-left {
+.controls-section {
   display: flex;
   align-items: center;
+  gap: 24px;
 }
 
+/* Transport Controls */
 .transport-controls {
   display: flex;
   gap: 8px;
 }
 
-/* Right Section - Position and BPM */
-.transport-right {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-/* Position Display */
-.position-display {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-/* BPM Controls */
-.bpm-controls {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.bmp-label {
-  font-weight: 600;
-  color: #6c757d;
-}
-
-.bpm-slider {
-  width: 120px;
-}
-
-/* Transport Buttons - Mockup Style */
 .transport-btn {
   border: none;
   border-radius: 12px;
@@ -58,65 +31,186 @@
   font-weight: 600;
   font-size: 14px;
   cursor: pointer;
-  transition: all 0.15s ease;
+  transition: all var(--transition-fast, 0.15s ease);
   display: inline-flex;
   align-items: center;
   justify-content: center;
   gap: 6px;
-  box-shadow: 0 0.125rem 0.25rem rgba(0, 0, 0, 0.075);
+  box-shadow: var(--shadow-sm, 0 0.125rem 0.25rem rgba(0, 0, 0, 0.075));
   color: white;
 }
 
-.transport-btn:hover {
+.transport-btn--base {
+  background: #6c757d;
+  color: white;
+}
+
+.transport-btn--base:hover {
   transform: translateY(-1px);
-  box-shadow: 0 0.5rem 1rem rgba(0, 0, 0, 0.15);
+  box-shadow: var(--shadow-md, 0 0.5rem 1rem rgba(0, 0, 0, 0.15));
 }
 
-/* Play/Pause Button */
-.transport-btn--play-pause {
+.transport-btn--play {
   background: #28a745;
-  min-width: 80px;
+  width: 95px; /* Fixed width to prevent button movement */
 }
 
-.transport-btn--play-pause:hover {
+.transport-btn--play:hover {
   background: #218838;
+  transform: translateY(-1px);
+  box-shadow: var(--shadow-md, 0 0.5rem 1rem rgba(0, 0, 0, 0.15));
 }
 
-/* Stop Button */
+.transport-btn--pause {
+  background: #ffc107;
+  color: #212529;
+  width: 95px; /* Fixed width to prevent button movement */
+}
+
+.transport-btn--pause:hover {
+  background: #e0a800;
+  transform: translateY(-1px);
+  box-shadow: var(--shadow-md, 0 0.5rem 1rem rgba(0, 0, 0, 0.15));
+}
+
 .transport-btn--stop {
   background: #dc3545;
 }
 
 .transport-btn--stop:hover {
   background: #c82333;
+  transform: translateY(-1px);
+  box-shadow: var(--shadow-md, 0 0.5rem 1rem rgba(0, 0, 0, 0.15));
 }
 
-/* Loop Button */
 .transport-btn--loop {
   background: #6f42c1;
 }
 
 .transport-btn--loop:hover {
   background: #5a32a3;
+  transform: translateY(-1px);
+  box-shadow: var(--shadow-md, 0 0.5rem 1rem rgba(0, 0, 0, 0.15));
 }
 
-/* Transport Badges - replaces Bootstrap badges */
-.transport-badge {
-  padding: 4px 8px;
-  border-radius: 4px;
-  font-size: 12px;
+/* Snap Control */
+.snap-control {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.snap-checkbox {
+  margin: 0;
+}
+
+.snap-label {
+  margin: 0;
+  font-weight: 500;
+  color: #495057;
+  cursor: pointer;
+}
+
+/* BPM Control */
+.bpm-control {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  background: rgba(91, 192, 235, 0.1);
+  padding: 12px 16px;
+  border-radius: 12px;
+  border: 1px solid rgba(91, 192, 235, 0.2);
+}
+
+.bpm-label {
   font-weight: 600;
+  color: #5bc0eb;
+  margin: 0;
+  font-size: 14px;
+}
+
+.bmp-slider {
+  width: 120px;
+  margin: 0;
+  accent-color: #5bc0eb;
+}
+
+.bpm-value {
+  background: #5bc0eb;
   color: white;
+  padding: 4px 8px;
+  border-radius: 8px;
+  font-weight: 600;
+  font-size: 14px;
+  min-width: 48px; /* Fixed width to prevent size changes */
+  text-align: center;
 }
 
-.transport-badge--beat {
-  background: #17a2b8;
+/* Measure Controls */
+.measure-controls {
+  display: flex;
+  align-items: center;
+  gap: 8px;
 }
 
-.transport-badge--tick {
-  background: #6c757d;
+.measure-label {
+  color: var(--text-muted, #adb5bd);
+  font-weight: bold;
+  font-size: 14px;
+  margin: 0;
 }
 
-.transport-badge--bpm {
-  background: #007bff;
+.measure-btn {
+  padding: 4px 8px;
+  border: 1px solid var(--border-color, #dee2e6);
+  background: white;
+  border-radius: 6px;
+  cursor: pointer;
+  font-size: 14px;
+  font-weight: 500;
+  color: var(--text-primary, #333);
+  transition: all 0.15s ease;
+  min-width: 28px;
+  height: 28px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.measure-btn:hover {
+  background: #f8f9fa;
+  border-color: #adb5bd;
+}
+
+.measure-btn:active {
+  background: #e9ecef;
+}
+
+.measure-btn:disabled {
+  background: #f8f9fa;
+  border-color: #e9ecef;
+  color: #adb5bd;
+  cursor: not-allowed;
+}
+
+.measure-count {
+  background: var(--text-secondary, #6c757d);
+  color: white;
+  padding: 4px 8px;
+  border-radius: 6px;
+  font-weight: 600;
+  font-size: 14px;
+  min-width: 28px;
+  height: 28px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+/* Focus states for accessibility */
+.transport-btn:focus,
+.snap-checkbox:focus,
+.bmp-slider:focus {
+  outline: 2px solid var(--playhead-color, #007bff);
+  outline-offset: 2px;
 }
